Rename folder components to shared
