+++ src/app/pages/challenges/challenge-list/challenge-list-filters-values.ts
@@ -1,5 +1,5 @@
-import { ButtonToggleFilterValue } from 'src/app/components/filters/button-toggle-filter/button-toggle-filter-value';
-import { FilterValue } from 'src/app/components/filters/filter-value.model';
+import { ButtonToggleFilterValue } from 'src/app/shared/filters/button-toggle-filter/button-toggle-filter-value';
+import { FilterValue } from 'src/app/shared/filters/filter-value.model';
 
 export const orderByFilterValues: FilterValue[] = [
   {
