+++ src/app/pages/homepage/homepage.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { AvatarModule, FooterModule } from '@sage-bionetworks/sage-angular';
 import { HomepageComponent } from './homepage.component';
-import { MaterialModule } from 'src/app/components/material/material.module';
+import { MaterialModule } from 'src/app/shared/material/material.module';
 
 const routes: Routes = [
   { path: '', component: HomepageComponent }
