+++ src/app/pages/organizations/organization-list/organization-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { PageTitleService } from 'src/app/components/page-title/page-title.service';
+import { PageTitleService } from 'src/app/shared/page-title/page-title.service';
 
 @Component({
   selector: 'rocc-organization-list',
