+++ src/app/pages/organizations/organizations.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FooterModule } from '@sage-bionetworks/sage-angular';
-import { MaterialModule } from 'src/app/components/material/material.module';
+import { MaterialModule } from 'src/app/shared/material/material.module';
 import { OrganizationsComponent } from './organizations.component';
 import { OrganizationsRoutingModule } from './organizations-routing.module';
 import { OrganizationListComponent } from './organization-list/organization-list.component';
