+++ src/app/pages/page-not-found/page-not-found.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { FooterModule } from '@sage-bionetworks/sage-angular';
-import { MaterialModule } from 'src/app/components/material/material.module';
+import { MaterialModule } from 'src/app/shared/material/material.module';
 import { PageNotFoundComponent } from './page-not-found.component';
 
 const routes: Routes = [
