+++ src/app/shared/challenge-preview/challenge-preview.component.html
