+++ src/app/shared/challenge-preview/challenge-preview.component.spec.ts
