+++ src/app/shared/filters/checkbox-filter/checkbox-filter.scss
