+++ src/app/shared/filters/search-filter/search-filter.scss
