+++ src/app/shared/filters/select-filter/select-filter.scss
